Split info into different files
